Tournaments db crud
