+++ src/tournaments/tournaments.module.ts
@@ -1,9 +1,19 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { TournamentsService } from './tournaments.service';
+import { TournamentsMemoryService } from './tournaments-memory.service';
+import { TournamentsDbService } from './tournaments-db.service';
 import { TournamentsController } from './tournaments.controller';
+import { Tournament } from './entities/tournament.entity';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([Tournament])],
   controllers: [TournamentsController],
-  providers: [TournamentsService],
+    providers: [
+      {
+        provide: 'TournamentsService',
+        useClass: TournamentsDbService
+      },
+    ]
 })
 export class TournamentsModule {}
